EXPRES

EchelleInstruments.EXPRES — Module

Module providing types and traits and customized functions for the EXPRES Spectrograph.

EXPRES Blog

"EXPRES: a next generation RV spectrograph in the search for earth-like worlds" (Jurgenson et al. 2016)

Types

EchelleInstruments.EXPRES.EXPRES1D — Type

Trait for 1D Extracted spectra from EXPRES

EchelleInstruments.EXPRES.EXPRES2D — Type

Trait for 2D Extracted spectra from EXPRES

EchelleInstruments.EXPRES.AnyEXPRES — Type

Trait to specify any 1D or 2D Extracted spectra from EXPRES

Functions

EchelleInstruments.EXPRES.continuum_normalize_spectra! — Method

Normalize each spectrum based on continuum model from FITS files.

EchelleInstruments.EXPRES.continuum_normalize_spectrum! — Method

Normalize spectrum based on continuum model from FITS file.

EchelleInstruments.EXPRES.filter_line_list — Method

filter_line_list( linelist, inst ; λmin, λmax ) Return DataFrame based on linelist, but filtered to only include lines between λmin and λmax. Defaults for λmin and λmax specified are in traits for EXPRES.

EchelleInstruments.EXPRES.find_worst_telluric_in_each_chunk — Method

find_worst_telluric_in_each_chunk( chunk_list_timeseries, spectra ) Return an array of the worst (smallest) telluric for each chunk at any time for the provided ChunkListTimeseries.

RvSpectMLBase.InstrumentsCommon.make_clean_line_list_from_tellurics — Method

make_clean_line_list_from_tellurics(line_list, expres_data; Δv_to_avoid_tellurics ) Returns a new line list that excludes lines with telluric contamination. Inputs:

line_list: Dataframe containing field lambda

expres_data: Array of spectra

Δv_to_avoid_tellurics: in m/s

Outputs:

line_list_without_tellurics: DataFrame with fields: lambda, weight, lambda_lo, and lambda_hi.

Warning: Currently, assumes a tellurics value in metadata for each spectra, such as is provided by EXPRES.

Files

Internals

EchelleInstruments.EXPRES.add_metadata_from_fits! — Method

add_metadata_from_fits!(df, filename) Return updated dataframe after adding metadata from FITS file header.

EchelleInstruments.EXPRES.blaze_normalize_spectra! — Method

Normalize each spectrum based on blaze model (also tries to include black body function) from FITS files.

EchelleInstruments.EXPRES.blaze_normalize_spectrum! — Method

Normalize spectrum based on blaze model (also tries to include black body function) from FITS file.

EchelleInstruments.EXPRES.jd2mjd — Method

Return modified Julian date based on input Julian date

EchelleInstruments.EXPRES.read_data — Function

Read EXPRES data from FITS file, and return in a Spectra2DBasic object.

EchelleInstruments.EXPRES.read_data — Method

Read EXPRES data from FITS a file using Filename from a DataFrameRow. Adds other columns from DataFrameRow as metadata.

EchelleInstruments.EXPRES.read_data_only — Method

Read only EXPRES data from FITS file, and leave metadata empty.

EchelleInstruments.EXPRES.read_data_only — Method

Read only EXPRES data from FITS file, and leave metadata empty.

EchelleInstruments.EXPRES.read_metadata — Method

Create Dict containing filename and default metadata from FITS file headers.

EchelleInstruments.EXPRES.read_metadata — Method

Create Dict containing filename and default metadata from FITS file headers.

EchelleInstruments.make_manifest — Method

Create Dataframe containing filenames and key data for all files neid*.fits in directory

Index

EchelleInstruments.EXPRES

EchelleInstruments.EchelleInstruments

EchelleInstruments.NEID

EchelleInstruments.EXPRES.AnyEXPRES

EchelleInstruments.EXPRES.EXPRES1D

EchelleInstruments.EXPRES.EXPRES2D

EchelleInstruments.NEID.AnyNEID

EchelleInstruments.NEID.NEID1D

EchelleInstruments.NEID.NEID2D

EchelleInstruments.EXPRES.add_metadata_from_fits!

EchelleInstruments.EXPRES.blaze_normalize_spectra!

EchelleInstruments.EXPRES.blaze_normalize_spectrum!

EchelleInstruments.EXPRES.continuum_normalize_spectra!

EchelleInstruments.EXPRES.continuum_normalize_spectrum!

EchelleInstruments.EXPRES.filter_line_list

EchelleInstruments.EXPRES.find_worst_telluric_in_each_chunk

EchelleInstruments.EXPRES.jd2mjd

EchelleInstruments.EXPRES.read_data

EchelleInstruments.EXPRES.read_data

EchelleInstruments.EXPRES.read_data_only

EchelleInstruments.EXPRES.read_data_only

EchelleInstruments.EXPRES.read_metadata

EchelleInstruments.EXPRES.read_metadata

EchelleInstruments.NEID.make_manifest

EchelleInstruments.NEID.read_activity_table

EchelleInstruments.NEID.read_blaze

EchelleInstruments.NEID.read_cal_data

EchelleInstruments.NEID.read_data

EchelleInstruments.NEID.read_differential_extinctions!

EchelleInstruments.NEID.read_exposure_meter

EchelleInstruments.NEID.read_metadata

EchelleInstruments.NEID.read_solar_normalization

EchelleInstruments.NEID.read_telluric_ranges

EchelleInstruments.calc_complement_wavelength_ranges

EchelleInstruments.find_ranges_with_tellurics

EchelleInstruments.make_manifest

EchelleInstruments.make_manifest

EchelleInstruments.make_manifest

EchelleInstruments.make_manifest

EchelleInstruments.make_ranges_without_tellurics

EchelleInstruments.merge_sorted_wavelength_ranges

EchelleInstruments.read_fits_header

EchelleInstruments.read_header

EchelleInstruments.read_header

EchelleInstruments.read_manifest

EchelleInstruments.read_metradata_from_fits

EchelleInstruments.write_manifest

RvSpectMLBase.InstrumentsCommon.make_clean_line_list_from_tellurics

RvSpectMLBase.InstrumentsCommon.make_clean_line_list_from_tellurics