NEID

EchelleInstruments.NEID — Module

Module providing types and traits and customized functions for the NEID Spectrograph.

NEID blog

NEID

Types

Functions

Internals

Index

Types

EchelleInstruments.NEID.NEID1D — Type

Trait for 1D Extracted spectra from NEID

EchelleInstruments.NEID.NEID2D — Type

Trait for 2D Extracted spectra from NEID

EchelleInstruments.NEID.AnyNEID — Type

Trait to specify any 1D or 2D Extracted spectra from NEID

Functions

RvSpectMLBase.InstrumentsCommon.make_clean_line_list_from_tellurics — Method

make_clean_line_list_from_tellurics(line_list, neid_data; Δv_to_avoid_tellurics ) Returns a new line list that excludes lines with telluric contamination. Inputs:

line_list: Dataframe containing field lambda

- neid_data: Array of spectra

Δv_to_avoid_tellurics: in m/s

v_center_to_avoid_tellurics : in m/s

tellurics_filename: filename with wavelength ranges affected by tellurics

Outputs:

line_list_without_tellurics: DataFrame with fields: lambda, weight, lambda_lo, and lambda_hi.

Warning: Currently, reads in a list of wavelength ranges with tellurics and this file was created based on tellurics metadata in EXPRES data for HR 101501.

Internals

EchelleInstruments.NEID.make_manifest — Method

Create Dataframe containing filenames and key data for all files neid*.fits in directory

EchelleInstruments.NEID.read_cal_data — Method

Read calibrator data

EchelleInstruments.NEID.read_data — Function

Read NEID data from FITS file, and return in a Spectra2DBasic object.

EchelleInstruments.NEID.read_differential_extinctions! — Function

Read space delimited file with differential extinction corrections, interpolate to bjd's in df and insert into df[:,diffextrv].

EchelleInstruments.NEID.read_metadata — Method

Create Dict containing filename and default metadata from file.

EchelleInstruments.NEID.read_solar_normalization — Function

Read normalization function for approximate solar continuum+blaze correction TODO: Figure out mapping between these orders and what's stored in the FITS file before using.

EchelleInstruments.NEID.read_telluric_ranges — Method

readtelluricranges( filename ) Return DataFrame (keys lambdalo and lambdahi) with wavelength ranges to be avoided as containing tellurics based on provided CSV file. Assumes path is included in filename.

Index

EchelleInstruments.EXPRES

EchelleInstruments.EchelleInstruments

EchelleInstruments.NEID

EchelleInstruments.EXPRES.AnyEXPRES

EchelleInstruments.EXPRES.EXPRES1D

EchelleInstruments.EXPRES.EXPRES2D

EchelleInstruments.NEID.AnyNEID

EchelleInstruments.NEID.NEID1D

EchelleInstruments.NEID.NEID2D

EchelleInstruments.EXPRES.add_metadata_from_fits!

EchelleInstruments.EXPRES.blaze_normalize_spectra!

EchelleInstruments.EXPRES.blaze_normalize_spectrum!

EchelleInstruments.EXPRES.continuum_normalize_spectra!

EchelleInstruments.EXPRES.continuum_normalize_spectrum!

EchelleInstruments.EXPRES.filter_line_list

EchelleInstruments.EXPRES.find_worst_telluric_in_each_chunk

EchelleInstruments.EXPRES.jd2mjd

EchelleInstruments.EXPRES.make_manifest

EchelleInstruments.EXPRES.read_data

EchelleInstruments.EXPRES.read_data

EchelleInstruments.EXPRES.read_data_only

EchelleInstruments.EXPRES.read_data_only

EchelleInstruments.EXPRES.read_metadata

EchelleInstruments.EXPRES.read_metadata

EchelleInstruments.NEID.make_manifest

EchelleInstruments.NEID.read_cal_data

EchelleInstruments.NEID.read_data

EchelleInstruments.NEID.read_differential_extinctions!

EchelleInstruments.NEID.read_metadata

EchelleInstruments.NEID.read_solar_normalization

EchelleInstruments.NEID.read_telluric_ranges

EchelleInstruments.calc_complement_wavelength_ranges

EchelleInstruments.find_ranges_with_tellurics

EchelleInstruments.make_manifest

EchelleInstruments.make_manifest

EchelleInstruments.make_manifest

EchelleInstruments.make_ranges_without_tellurics

EchelleInstruments.merge_sorted_wavelength_ranges

EchelleInstruments.read_fits_header

EchelleInstruments.read_header

EchelleInstruments.read_header

EchelleInstruments.read_manifest

EchelleInstruments.read_metradata_from_fits

EchelleInstruments.write_manifest

RvSpectMLBase.InstrumentsCommon.make_clean_line_list_from_tellurics

RvSpectMLBase.InstrumentsCommon.make_clean_line_list_from_tellurics